Down with the king, up with people and locations!
Summary:
Starting to wonder whether this was the right project to fork. But why not keep
going with this path? Instead of updating the one king (or the "owner
of the website"), this changeset allows there to be two people (Marcia and
Kalvin) and you can update their respective locations.

Removed redis too, since we're not keeping track of the last update time.

Test Plan:
Open up localhost:1234 in two tabs
See Marcia and Kalvin's locations
Update one, update another
Marvel!